Update all the samples.
(except clojure + scalatra, because those are broken, will fixed with later master merge.)

+++ samples/client/petstore/tizen/src/User.h
@@ -1,23 +1,7 @@
 /*
- * Copyright (c) 2016 Samsung Electronics Co., Ltd All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
  * User.h
  *
- * 
+ * A User who is purchasing from the pet store
  */
 
 #ifndef _User_H_
@@ -31,7 +15,7 @@ namespace Tizen {
 namespace ArtikCloud {
 
 
-/*! \brief 
+/*! \brief A User who is purchasing from the pet store
  *
  */
 
@@ -41,7 +25,7 @@ public:
 	 */
 	User();
 	User(char* str);
-	
+
 	/*! \brief Destructor.
 	 */
 	virtual ~User();
@@ -56,11 +40,11 @@ public:
 
 	/*! \brief Get 
 	 */
-	long getId();
+	long long getId();
 
 	/*! \brief Set 
 	 */
-	void setId(long  id);
+	void setId(long long  id);
 	/*! \brief Get 
 	 */
 	std::string getUsername();
@@ -112,7 +96,7 @@ public:
 	void setUserStatus(int  userStatus);
 
 private:
-	long id;
+	long long id;
 	std::string username;
 	std::string firstName;
 	std::string lastName;
@@ -122,7 +106,7 @@ private:
 	int userStatus;
 	void __init();
 	void __cleanup();
-	
+
 };
 }
 }
